feat(wg): split WireGuard implementation so that in-kernel and userspace implementations can coexist
