+++ pkg/wg/userspace/userspace-wg.go
@@ -1,0 +1,1 @@
+package userspacewg
